correct mistake

+++ src/solver.jl
@@ -115,7 +115,7 @@ function lsoda(f::Function, y0::Vector{Float64}, tspan::Vector{Float64}; userdat
   return ctx, yres
 end
 
-function lsoda_evolve!(ctx::lsoda_context_t,y::Vector{Float64},tspan::Vector{Float})
+function lsoda_evolve!(ctx::lsoda_context_t,y::Vector{Float64},tspan::Vector{Float64})
 	t    = Array{Float64}(1)
 	tout = Array{Float64}(1)
 	t[1] = tspan[1]
